requirements.txt
# Core dependencies
requests>=2.31.0
pyyaml>=6.0.1
python-dotenv>=1.0.0

# LLM providers
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0

# UI Framework
tkinter-tooltip>=2.0.0
Pillow>=10.0.0

# Utilities
rich>=13.0.0
click>=8.1.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
